+++ extensions/ipynb/package.json
@@ -9,7 +9,6 @@
 		"vscode": "^1.57.0"
 	},
 	"enabledApiProposals": [
-		"notebookEditor",
     "notebookWorkspaceEdit"
 	],
 	"activationEvents": [
