+++ extensions/vscode-api-tests/package.json
@@ -24,7 +24,6 @@
     "notebookControllerKind",
     "notebookDebugOptions",
     "notebookDeprecated",
-    "notebookEditor",
     "notebookEditorDecorationType",
     "notebookEditorEdit",
     "notebookLiveShare",
