+++ extensions/vscode-notebook-tests/package.json
@@ -15,7 +15,6 @@
     "notebookControllerKind",
     "notebookDebugOptions",
     "notebookDeprecated",
-    "notebookEditor",
     "notebookEditorDecorationType",
     "notebookLiveShare",
     "notebookMessaging",
